some necessary technical debt to be refactored later

+++ app/src/main/java/cin/ufpe/br/microbit_car_assist/domain/entities/AccelerometerData.kt
@@ -5,4 +5,4 @@ package cin.ufpe.br.microbit_car_assist.domain.entities
  * at 05/23/2018 5:58 PM
  */
 
-data class AccelerometerData(val x: Short, val y: Short, val z: Short, val pitch: Double, val roll: Double, var isHole: Boolean = false)
+data class AccelerometerData(val x: Float, val y: Float, val z: Float, val pitch: Double, val roll: Double, var isHole: Boolean = false)
